% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addANTI}
\alias{addANTI}
\title{An Anti MGF function}
\usage{
addANTI(mgf)
}
\arguments{
\item{mgf}{The mgf data.table object}
}
\description{
This function recieves an mgf data.table object and adds the decoy specta to it. Each spectrum gets an ANTI spectum which is used as a decoy for clustering algorithm.
}
\examples{
mgf_with_decoys<-addANTI(mgf1)
}
\keyword{anti}
