% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addFDRtoLibrary}
\alias{addFDRtoLibrary}
\title{A get FDR table function
This function recieves a spectral library created be CreateLIBfromMGF function and adds FDR to it. FDR is calculated using the getFDRtable function.}
\usage{
addFDRtoLibrary(library)
}
\arguments{
\item{library}{A spectral library}
}
\description{
A get FDR table function
This function recieves a spectral library created be CreateLIBfromMGF function and adds FDR to it. FDR is calculated using the getFDRtable function.
}
\examples{
lib1<-addFDRtoLibrary(lib1)
}
\keyword{FDR}
