% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expendLibIDs}
\alias{expendLibIDs}
\title{expend Lib IDs  function
This function recieves a spectral library with added Ids and expends them to the containing clusters.}
\usage{
expendLibIDs(library)
}
\arguments{
\item{library}{A spectral library}
}
\description{
expend Lib IDs  function
This function recieves a spectral library with added Ids and expends them to the containing clusters.
}
\examples{
lib1<-expendLibIDs(lib1)
}
\keyword{Expend}
