% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getFullMgf}
\alias{getFullMgf}
\title{get Full MGF from Lib function
This function recieves a spectral library and returns full mgf data.table object of all the spectra wiyjin the library.}
\usage{
getFullMgf(library)
}
\arguments{
\item{library}{A spectral library}
}
\description{
get Full MGF from Lib function
This function recieves a spectral library and returns full mgf data.table object of all the spectra wiyjin the library.
}
\examples{
mgf1<-getFullMgf(lib1)
}
\keyword{full}
