% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getLastMgfFromLib}
\alias{getLastMgfFromLib}
\title{get last MGF from Lib function
This function recieves a spectral library and returns an mgf data.table object of the mgf last added to the library.}
\usage{
getLastMgfFromLib(library)
}
\arguments{
\item{library}{A spectral library}
}
\description{
get last MGF from Lib function
This function recieves a spectral library and returns an mgf data.table object of the mgf last added to the library.
}
\examples{
mgf1<-getLastMgfFromLib(lib1)
}
\keyword{last}
