% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{noise_estimation}
\alias{noise_estimation}
\title{A noise estimation function}
\usage{
noise_estimation(int)
}
\arguments{
\item{int}{A list of peak intensities form MSMS spectrum.}
}
\description{
This function estimates noise level for a list of peak intensities. Supports Centroin and CWT peak pickings.
}
\examples{
noise <- noise_estimation(c(1,2,3,4,5))
}
\keyword{noise}
