% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Prepare_MGF}
\alias{Prepare_MGF}
\title{A Prepare mgf function
This function recieves a path to the .mgf file and returnes an MGF data.table object. Mgf is parsed using Parse_mgf and nise_estimation functions, and prepared to clustering.}
\usage{
Prepare_MGF(path, mode = "centroid")
}
\arguments{
\item{path}{A path to the .mgf file}

\item{mode}{A mode of mgf preparation. Can be PP for Protein PIlot type of mgf, and centroid for the centroid mgf type. Defaults to centroid.}
}
\description{
A Prepare mgf function
This function recieves a path to the .mgf file and returnes an MGF data.table object. Mgf is parsed using Parse_mgf and nise_estimation functions, and prepared to clustering.
}
\examples{
mgf1<-Prepare_MGF('c:/docs/mgf1.mgf')
}
\keyword{prepareMgf}
